+++ zeeguu_web/account/api/account_management.py
@@ -2,7 +2,7 @@ import flask
 import requests
 import zeeguu
 
-from zeeguu_web.account.api import APIConnectionError, API_POST
+from zeeguu_web.account.api import API
 
 
 class AccountManagement:
@@ -14,11 +14,11 @@ class AccountManagement:
     @classmethod
     def create_account(cls, email, username, password, learning_language, native_language):
         api_address = cls.api_create_user + email
-        resp = API_POST(api_address, payload = {"password": password, "username":username})
+        resp = API.post(api_address, payload = {"password": password, "username":username})
         session = int(resp.text)
             # What to do when the server crashes here?
         native = cls.api_native + native_language.code
         learned = cls.api_learned + learning_language.code
-        resp = API_POST(native, session_needed=True, session=session)
-        resp = API_POST(learned, session_needed=True, session=session)
+        resp = API.post(native, session_needed=True, session=session)
+        resp = API.post(learned, session_needed=True, session=session)
         return session
