+++ zeeguu_web/account/api/api_exceptions.py
@@ -1,0 +1,5 @@
+class APIConnectionError(Exception):
+    def __init__(self, code):
+        super()
+        self.status_code = code
+
