+++ zeeguu_web/account/api/session_management.py
@@ -2,7 +2,7 @@ import flask
 import requests
 import zeeguu
 
-from zeeguu_web.account.api import API_GET, API_POST
+from zeeguu_web.account.api.API import API
 
 
 class SessionManagement:
@@ -14,15 +14,15 @@ class SessionManagement:
     @classmethod
     def login(cls, email, password):
         url = cls.api_login + email
-        resp = API_POST(url, payload={"password": password})
+        resp = API.post(url, payload={"password": password})
         return resp.text
 
     @classmethod
     def logout(cls):
-        API_GET(cls.api_logout, session_needed=True)
+        API.get(cls.api_logout, session_needed=True)
 
     @classmethod
     def validate(cls):
-        API_GET(cls.api_validate, session_needed=True)
+        API.get(cls.api_validate, session_needed=True)
